scrapping_app/requirements.txt
selenium
pandas
numpy
webdriver-manager
pymongo==4.5.0
psycopg2-binary==2.9.7
sqlalchemy
